requirements.txt
flask==3.0.2
flask-cors==4.0.0
PyMuPDF==1.25.0
python-pptx==0.6.23
Werkzeug==3.0.1
pdf2image>=1.16.3
Pillow>=10.2.0
click>=8.1.7
colorama>=0.4.6
itsdangerous>=2.1.2
Jinja2>=3.1.3
MarkupSafe>=2.1.5
packaging>=23.2
XlsxWriter>=3.1.9
lxml>=5.1.0 